+++ hub/build.sbt
@@ -1,8 +1,9 @@
+
 name := "hub"
 
 version := "0.1"
-
 scalaVersion := "2.13.2"
+mainClass in assembly := Some("eu.jrie.put.wti.bigstore.hub.HubService")
 
 libraryDependencies ++= Seq(
   // akka
@@ -14,3 +15,10 @@ libraryDependencies ++= Seq(
   // util
   "ch.qos.logback" % "logback-classic" % "1.2.3"
 )
+
+assemblyMergeStrategy in assembly := {
+  case PathList("META-INF", xs @ _*) => MergeStrategy.discard
+  case "reference.conf" => MergeStrategy.concat
+  case _ => MergeStrategy.first
+}
+
